✨ Add either<A, B> monadic type
This differs from outcome<T, M, F> (planned) which is for error
handling, variant<Ts...> (which requires more work and a different
approach to member access while also possibly generating worse code),
but is still under the "monadic" umbrella of vocabulary types.

Effectively, apex::either<A, B> is to std::variant as std::pair is to
std::tuple.

However, we have not yet added the full monadic interface (and_then,
transform, etc.) and that is required for more work.

✨ Add bad_either_access exception
✨ Add specialization_of<T, U> concept
✨ Add safely_invocable<F, Args...> concept
✨ Add safely_copy_assignable<T> concept
✨ Add safely_move_assignable<T> concept
✅ Add initial set of tests for either<A, B>