🐛 Fix incorrect inheritance cascade.
We CANNOT inherit constructors to get the same behavior for each type.
This means every custom error MUST take an apex::source_location.
Furthermore, the function_name() will be correct, but the line will be
from within the function. Care should be taken when relying on this
behavior, as it might differ in the future.

🐛 Fix compilation errors