🐛 Fix #24 and refactor template constraints
The template constraint was written a bit incorrectly, and could result
in unreadable error messages. This fixes that, as well as speeds up
compilation by removing the call to std::forward and replacing it with
static_cast to a forwarding reference.

✅ Add back_emplacer usage test
For some reason, this never made it into the repo when it was moved to netlify