📝 Add initial set of documentation files
Running `--target docs` is enough to get the target to build.

A lot of these entries are currently stubs as we need to move doxygen
comments over, write up better interfaces, clean up a few APIs, etc.

At some point I would like to get the workflow attached to this setup to
build/deploy to netlify, however the netlify build image does not have
cmake installed, and that's a requirement since our netlify/cmake
library is where the central set of documentation settings are stored at
the moment