♻ Massive refactor of apex::either<A, B>
This refactor improves compile times, fixes a few constraints to be more
correct and... constrained, as well as permits references as members. We
do not currently have a way to `emplace` a reference into an either, and
we might not want to support this just yet, as it has the same issue as
std::optional's history regarding assign-through or rebind.

This refactor also moved a bunch of detail template specialization out
of the detail namespace and into the main file itself. While this makes
the core file a bit more verbose the important bits are all in one
place (mostly)

More work is, of course, needed regarding transformation calls, making
either work with std::get (and thus structured bindings), and
std::variant-like interfaces.

We still need to decide if mirroring the variant API or the optional API
is more important. Additionally, we need to decide if the usage of
`in_place_index` and `in_place_type` are actually what we want and not
an explicit `left`/`right` dichotomy as sentinel types, *with* the
option of using the `in_place_XXX` declarations as the

✨ Add <apex/core/types.hpp>

Although it was part of the prelude, the types header has been useful to
keep small, and having it effectively "alias" to the prelude internal
header has reduced some compile time overhead. That said, we might want
to revisit the use of the prelude as it is beginning to negatively
affect compile time performance for small components that might not be
used with other bits and pieces.

✨ Add lvalue_reference concept

This is currently poorly named and surprisingly, we tend to need *both*
the `lvalue_reference` and `not lvalue_reference` form. We also *never*
need the rvalue in practice, as `auto&&` means that only the `auto` is
counted and *then* the rvalue reference is applied. Because of this, we
might make it `apex::reference` and `apex::object` as concepts. (Where
`apex::object` is `not apex::reference and ::std::is_object_v`)

📝 Mock out netlify deployment step